+++ CHANGELOG.md
@@ -1,6 +1,16 @@
-##2015-08-23 - Release 0.1.0
-###Summary
+## 2016-12-25 - Release 0.2.0
+### Summary
+This release adds support to Ubuntu Xenial, fixes typo and a bad URL and  adds a lot of test improvements.
+
+#### Features
+- Added support to Ubuntu 16.04.
+
+### Bugfixes
+- Fixed `all_devel` and `doc` parameters usage.
+
+## 2015-08-23 - Release 0.1.0
+### Summary
 This is the initial release.
 
-####Features
+#### Features
 - Install Boost for Debian, Ubuntu, RedHat and CentOS